datamonkey/requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
PyYAML
SQLAlchemy
Werkzeug==0.9.4
boto==2.35.1
clay-flask
iso8601==0.1.8
itsdangerous==0.23
oursql
psycopg2
pytz==2013.9
redshift-sqlalchemy==0.4.1
requests==2.2.0
simplejson==3.3.2
simple-salesforce==0.63.0
ua-parser==0.3.4
wsgiref==0.1.2
pymongo
mrjob
TwitterAPI
python-instagram
kafka-python
cassandra-driver
